requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-unfold==0.30.0
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.0
whitenoise==6.7.0
